gateway/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
packaging==23.0
pika==1.3.1
Pillow==9.4.0
pycodestyle==2.10.0
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
